Add constant file and refactor
